requirements.txt
# Kafka
kafka-python==2.0.2

# async http requests package
aiohttp==3.8.4
aiokafka==0.8.0
